[feat] 자동차 InputVIew 전략 패턴 구현
