+++ src/main/java/racingcar/strategy/InputStrategy.java
@@ -1,0 +1,10 @@
+package racingcar.strategy;
+
+import racingcar.game.RacingInfomation;
+
+import java.util.Scanner;
+
+@FunctionalInterface
+public interface InputStrategy {
+    RacingInfomation question(Scanner scanner);
+}
